fix: feature/starbox/mvvm -> develop merge conflict 해결
