--- app/src/main/java/com/sainfe/todowithspot/AllActivity.java
@@ -1,29 +1,0 @@
-package com.sainfe.todowithspot;
-
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.constraintlayout.widget.ConstraintLayout;
-
-import android.annotation.SuppressLint;
-import android.content.Intent;
-import android.os.Bundle;
-
-public class AllActivity extends AppCompatActivity {
-
-    @SuppressLint("ClickableViewAccessibility")
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_all);
-
-        ConstraintLayout all_layout = findViewById(R.id.all_layout);
-        all_layout.setOnTouchListener(new OnSwipeTouchListener(this) {
-            @Override
-            public void onSwipeRight() {
-                super.onSwipeRight();
-                Intent swipeRightIntent = new Intent(getApplicationContext(), MainActivity.class);
-                startActivity(swipeRightIntent);
-                overridePendingTransition(R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
-            }
-        });
-    }
-}
