+++ app/src/main/java/com/sainfe/todowithspot/view/all/AllActivity.java
@@ -3,23 +3,37 @@ package com.sainfe.todowithspot.view.all;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.databinding.DataBindingUtil;
 
+import android.annotation.SuppressLint;
+import android.content.Intent;
 import android.os.Bundle;
 
 import com.sainfe.todowithspot.R;
 import com.sainfe.todowithspot.databinding.ActivityAllBinding;
+import com.sainfe.todowithspot.view.today.OnSwipeTouchListener;
+import com.sainfe.todowithspot.view.today.TodayActivity;
 import com.sainfe.todowithspot.viewmodel.all.AllViewModel;
-import com.sainfe.todowithspot.viewmodel.today.TodayViewModel;
 
 public class AllActivity extends AppCompatActivity {
 
     ActivityAllBinding binding;
 
+    @SuppressLint("ClickableViewAccessibility")
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_all);
         binding = DataBindingUtil.setContentView(this, R.layout.activity_all);
-        binding.setViewmodel(new AllViewModel());
+        binding.setViewModel(new AllViewModel());
         binding.executePendingBindings();
+
+        binding.allLayout.setOnTouchListener(new OnSwipeTouchListener(this) {
+            @Override
+            public void onSwipeRight() {
+                super.onSwipeRight();
+                Intent swipeRightIntent = new Intent(getApplicationContext(), TodayActivity.class);
+                startActivity(swipeRightIntent);
+                overridePendingTransition(R.anim.anim_slide_in_left, R.anim.anim_slide_out_right);
+            }
+        });
     }
 }
