+++ app/src/main/res/layout/activity_all.xml
@@ -6,11 +6,12 @@
     <data>
 
         <variable
-            name="viewmodel"
+            name="viewModel"
             type="com.sainfe.todowithspot.viewmodel.all.AllViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/all_layout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:context=".view.all.AllActivity">
