+++ app/src/main/res/layout/activity_today.xml
@@ -11,6 +11,7 @@
     </data>
 
     <LinearLayout
+        android:id="@+id/layout_today"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
